Integrate with Jenkins CI
Summary:
So we can track coverage over time and unit test results, and so that we can report failed builds in phab - we need to integrate the project with Jenkins. This diff adds pytest-cov so that junit test results and xml coverage reports can be generated.

@bypass-review because it doesn't need to be reviewed.

Test Plan: Should report coverage and build status in phabricator.

Reviewers: TomReichardt

Differential Revision: https://phab.adacs.org.au/D560