Add job upload
Summary:
To be able to upload jobs, we need to add code to interface with the job upload API. This diff adds two new methods to upload jobs, one takes an tar.gz archive of a job and uploads that, and the other takes a job output directory. The function that uploads the directory generates a temparory tar.gz file which is then uploaded.

Depends on https://phab.adacs.org.au/D590

Test Plan: TBD

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D589